refactor(product-list-service)
