+++ src/server/product-list/index.ts
@@ -1,3 +1,3 @@
-export type { IProductListService } from "./product-list.service.interface";
-export { PRODUCT_LIST_SERVICE_PROVIDER } from "./product-list.service.provider";
+export type { IProductListService } from "./product-list-service/product-list.service.interface";
+export { PRODUCT_LIST_SERVICE_PROVIDER } from "./product-list-service/product-list.service.provider";
 export { ProductListModule } from "./product-list.module";
