+++ src/server/product-list/product-list-service/product-list.service.interface.ts
@@ -1,0 +1,9 @@
+import type { IProductList } from "../../../product-list";
+
+interface IProductListService {
+  getData(): Promise<IProductList[]>;
+
+  saveData(data: IProductList[]): Promise<void>;
+}
+
+export type { IProductListService };
