--- src/server/product-list/product-list.service.interface.ts
@@ -1,7 +1,0 @@
-type ProductListStorageId = string;
-
-interface IProductListService {
-  initStorage(): Promise<ProductListStorageId>;
-}
-
-export type { IProductListService, ProductListStorageId };
